Use custom hook for expense form result
